LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T10-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token ')'. Expected identifier name as catch target.
S12.14_A16_T10

PASS Expected parsing failure

TEST COMPLETE
